Add files via upload
The main purpose of this project is to meet the requirements and specifications of the clients who always approach the ESS for there projects and for that reason we follow and maintain a proper sequence to complete the requirements. And as of now the Chef connect company which is a social media company has contacted the ESS and asked our company’s team to create a website where they can connect with world’s best chef’s and look for opportunities online.